Update README.md
Updated the readme to breathe a little more life into it, will fix the pdf generation bug as well as a bug with some of the entity stuff, and I'll upload the SRS document. Also, just throwing it out there: if anyone wants to turn this thing into a website and actually make it a product, that wouldn't be a terrible idea and I'm considering it personally.